- Refactors `core_data` module to `core-data`. - Updates the `settings.gradle.kts` file to include the new module name (`core-data`).
