+++ core-data/build.gradle.kts
@@ -1,6 +1,7 @@
 plugins {
     alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
+    alias(libs.plugins.kotlin.serialization)
 }
 
 android {
@@ -34,9 +35,37 @@ android {
 
 dependencies {
 
+    // Kotlin & Coroutines
+    implementation(libs.kotlin.stdlib)
+    implementation(libs.kotlinx.coroutines.core)
+    implementation(libs.kotlinx.coroutines.android)
+
+    // Retrofit & Networking
+    implementation(libs.retrofit)
+    implementation(libs.okhttp)
+    implementation(libs.logging.interceptor)
+
+    // JSON Parsing (Kotlinx Serialization)
+    implementation(libs.kotlinx.serialization.json)
+    // Converter do Retrofit para Kotlinx Serialization
+    implementation(libs.retrofit.kotlinx.serialization.converter)
+
+    // Dependency Injection (Koin)
+    // Koin Core features
+    implementation(libs.koin.core)
+    // Koin Android features
+    implementation(libs.koin.android)
+    // Koin for Jetpack Compose (se usar Compose futuramente)
+    // implementation("io.insert-koin:koin-androidx-compose:3.5.3")
+
+
+    // Koin Test features
+    testImplementation(libs.koin.test)
+    testImplementation(libs.koin.test.junit4)
+
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
-    implementation(libs.material)
+
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
