+++ core-data/src/androidTest/java/br/ftdev/core/data/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package br.ftdev.core_data
+package br.ftdev.core.data
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
