+++ core-data/src/main/AndroidManifest.xml
